+++ src/entities/profile/ui/profile-followers-info/profile-follower-info-block.tsx
@@ -2,8 +2,8 @@
 
 import { memo } from 'react'
 
-import { FollowersInfoHeader } from '@/entities/profile/ui/followers-info-header/followers-info-header'
-import { ProfileInfoDescription } from '@/entities/profile/ui/profile-info-description/profile-info-description'
+import { FollowersInfoHeader } from '@/entities/profile/ui/followers-info-header'
+import { ProfileInfoDescription } from '@/entities/profile/ui/profile-info-description'
 
 export const ProfileFollowerInfoBlock = memo(() => {
   return (
