+++ src/shared/ui/index.ts
@@ -19,8 +19,8 @@ export { Popover, PopoverContent, PopoverTrigger } from './popover'
 export { Recaptcha } from './recaptcha'
 export { Select, SelectBox, SelectContent, SelectItem, SelectTrigger } from './select'
 export { Sheet, SheetClose, SheetContent, SheetPortal, SheetTrigger } from './sheet'
+export { Skeleton } from './skeleton'
 export { TabContent, TabSwitcher } from './tabs'
 export { Text } from './text'
 export { ControlledTextarea, Textarea } from './textarea'
 export { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from './tooltip'
-export { UserAvatar } from './user-avatar'
